+++ src/server/spacetime/bindings/shard/deploy_mini_zone_reducer.ts
@@ -13,6 +13,6 @@ import {
 export default {
   clientTimeMs: __t.u64(),
   anchorCardId: __t.u32(),
-  targetMacroZone: __t.u32(),
+  targetMacroZone: __t.u64(),
   targetMicroZone: __t.u8(),
 };
