+++ src/server/spacetime/bindings/shard/souls_table.ts
@@ -15,7 +15,7 @@ export default __t.row({
   cardId: __t.u32().name("card_id"),
   ownerId: __t.u32().name("owner_id"),
   surface: __t.u8(),
-  macroZone: __t.u32().name("macro_zone"),
+  macroZone: __t.u64().name("macro_zone"),
   microZone: __t.u8().name("micro_zone"),
   microLocation: __t.u32().name("micro_location"),
   stats: __t.u32(),
