+++ src/server/spacetime/bindings/shard/zones_table.ts
@@ -14,7 +14,7 @@ export default __t.row({
   validAt: __t.u64().primaryKey().name("valid_at"),
   zoneId: __t.u32().name("zone_id"),
   surface: __t.u8(),
-  macroZone: __t.u32().name("macro_zone"),
+  macroZone: __t.u64().name("macro_zone"),
   packedDefinition: __t.u8().name("packed_definition"),
   ownerId: __t.u32().name("owner_id"),
   t0: __t.u64().name("t_0"),
